move ticker pkg to agent
